Change return type and input parameters to double

+++ src/main/java/com/example/springdocker/mathcalc/Calculator.java
@@ -17,7 +17,8 @@ public class Calculator {
         return a * b;
     }
 
-    public float divide(float a, float b) {
+    public double divide(double a, double b) {
+
         if(b==0) {
             throw new ArithmeticException("Can't divide by zero");
         } else {
